Correct non negative variance

+++ src/boot/ksdk1.1.0/pedometer.c
@@ -156,7 +156,7 @@ void read_acceleration_distribution(uint8_t N, int16_t * x_mean, int16_t * x_var
 	//int16_t z[N];
 
 	int16_t sum[3] = {0, 0, 0};
-	uint32_t sq_sum[3] = {0, 0, 0};
+	int16_t sq_sum[3] = {0, 0, 0};
 
 	acc_measurement measurement;
